requirements.txt
pydantic
requests
pypdf
python-docx
termcolor
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
streamlit
plotly
